Removed redundant code: singleton
