+++ dynamic_listings/lua/default/design_patterns/singleton.lua
@@ -1,15 +1,22 @@
-local function Singleton(superclass)
+local function LazySingleton(superclass)
     local o = {}
     o.__index = o
     setmetatable(o, superclass)
-    local instance = setmetatable({}, o)
-
-    o._instance = instance
 
     function o.new()
         -- We overwrite the constructor to something that returns
         -- the instance if already existing
+        if o._instance then
+            return o._instance
+        end
+
+        local instance = setmetatable({}, o)
+
+        -- Multi-threading: manage race conditions
+        -- ----- Critical region start -----
+        o._instance = instance
         return o._instance
+        -- ----- Critical region end -----
     end
 
     return o
