+++ dynamic_listings/pseudocode/default/design_patterns/singleton.txt
@@ -1,12 +1,19 @@
-class Singleton {
+class LazySingleton {
 
-    static Singleton INSTANCE = new Singleton();
+    static LazySingleton instance = null;
 
-    // This makes the constructor impossible to use outside of the class
+    // Block the constructor from working, we don't want to build more than one instance
     constructor() {}
 
     @staticmethod
-    function getInstance() -> Singleton{
-        return INSTANCE;
+    function getInstance() -> LazySingleton{
+        // Multi-threading: manage race conditions
+        // ----- Critical region start -----
+        if (instance == null) {
+            instance = new LazySingleton();
+        }
+        // ----- Critical region end -----
+
+        return instance;
     }
 }
