--- dynamic_listings/pseudocode/default/design_patterns/singleton_lazyload.txt
@@ -1,19 +1,0 @@
-class LazySingleton {
-
-    static LazySingleton instance = null;
-
-    // Block the constructor from working, we don't want to build more than one instance
-    constructor() {}
-
-    @staticmethod
-    function getInstance() -> LazySingleton{
-        // Multi-threading: manage race conditions
-        // ----- Critical region start -----
-        if (instance == null) {
-            instance = new LazySingleton();
-        }
-        // ----- Critical region end -----
-
-        return instance;
-    }
-}
